Add RabbitMQ package and update dependencies
Added a new package that establishes a connection with a RabbitMQ server, opens a channel and consumes messages from it. Accommodated the operation by adding "github.com/rabbitmq/amqp091-go" dependency.
The addition was necessary to ensure seamless communication with RabbitMQ server and efficient handling of messages. The dependency was updated in go.mod and go.sum to ensure the compatibility and proper functioning of the newly added package.